remove pool. keep allocator
