+++ TODO.md
@@ -19,3 +19,20 @@ doc.set('data.location', { latitude, longitude });
 let geopoint = doc.get('data.location') // GeoPoint
 geopoint.getProperties('latitude', 'longitude'); // 24.72504500749274, 58.74554729994484
 ```
+
+## Fastboot
+
+``` javascript
+const express = require('express');
+const fastbootMiddleware = require('fastboot-express-middleware');
+
+let app = express();
+
+app.use(express.static('../dist', { index: false }));
+
+app.get('/*', fastbootMiddleware('../dist'));
+
+app.listen(3000, () => {
+  console.log('FastBoot app listening on port 3000!');
+});
+```
