+++ addon/-private/firebase/instantiate.js
@@ -1,0 +1,3 @@
+import FirebaseInitializer from './initializer';
+
+export default (sender, identifier, opts) => new FirebaseInitializer(sender, identifier, opts);
